combined classes into one file
